Refactor feature strategy list to use DisabledInfo component
